requirements.txt
numpy~=1.26.4
resampy~=0.4.3
tensorflow~=2.16.1
tf_slim
six~=1.16.0
soundfile~=0.12.1
matplotlib~=3.8.4
pygame~=2.5.2
pyaudio~=0.2.14
pandas~=2.2.2
scikit-learn~=1.4.2
matplotlib
joblib~=1.4.0
pydub~=0.25.1
librosa~=0.10.2
pillow~=10.3.0
noisereduce~=3.0.2
scipy~=1.13.0